Split generally useful bits from AnnouncementViewController into separate files.
