+++ App/View Controllers/AnnouncementViewController.swift
@@ -115,8 +115,8 @@ final class AnnouncementViewController: ViewController {
         renderView.autoresizingMask = [.flexibleWidth, .flexibleHeight]
         view.insertSubview(renderView, at: 0)
 
-        renderView.registerJavaScriptMessage(RenderView.JavaScriptMessage.DidRender.self)
-        renderView.registerJavaScriptMessage(RenderView.JavaScriptMessage.DidTapAuthorHeader.self)
+        renderView.registerMessage(RenderView.BuiltInMessage.DidRender.self)
+        renderView.registerMessage(RenderView.BuiltInMessage.DidTapAuthorHeader.self)
 
         announcementObserver = ManagedObjectObserver(object: announcement, didChange: { [weak self] (change) in
             guard let sself = self else { return }
@@ -465,9 +465,9 @@ extension AnnouncementViewController: ComposeTextViewControllerDelegate {
 }
 
 extension AnnouncementViewController: RenderViewDelegate {
-    func didReceive(javaScriptMessage: RenderViewJavaScriptMessage, in view: RenderView) {
-        switch javaScriptMessage {
-        case is RenderView.JavaScriptMessage.DidRender:
+    func didReceive(message: RenderViewMessage, in view: RenderView) {
+        switch message {
+        case is RenderView.BuiltInMessage.DidRender:
             switch state {
             case .renderingFirstTime(let model), .rerendering(let model):
                 state = .rendered(model)
@@ -476,11 +476,11 @@ extension AnnouncementViewController: RenderViewDelegate {
                 Log.w("ignoring didRender in unexpected state \(state)")
             }
 
-        case let message as RenderView.JavaScriptMessage.DidTapAuthorHeader:
+        case let message as RenderView.BuiltInMessage.DidTapAuthorHeader:
             didTapAuthorHeaderInPost(at: message.postIndex, frame: message.frame)
 
         default:
-            Log.w("ignoring unexpected JavaScript message: \(type(of: javaScriptMessage).messageName)")
+            Log.w("ignoring unexpected JavaScript message: \(type(of: message).messageName)")
         }
     }
 
@@ -494,438 +494,5 @@ extension AnnouncementViewController: RenderViewDelegate {
         else {
             UIApplication.shared.openURL(url)
         }
-    }
-}
-
-final class RenderView: UIView {
-    weak var delegate: RenderViewDelegate?
-    var scrollView: UIScrollView { return webView.scrollView }
-
-    fileprivate var registeredJavaScriptMessages: [RenderViewJavaScriptMessage.Type] = []
-
-    fileprivate lazy var webView: WKWebView = {
-        let configuration = WKWebViewConfiguration()
-
-        for filename in ["Announcement.js", "RenderView.js"] {
-            let jsURL = Bundle(for: RenderView.self).url(forResource: filename, withExtension: nil) !! "Please include \(filename)"
-            let js = try! String(contentsOf: jsURL)
-            let userScript = WKUserScript(source: js, injectionTime: .atDocumentEnd, forMainFrameOnly: true)
-            configuration.userContentController.addUserScript(userScript)
-        }
-
-        let webView = WKWebView(frame: .zero, configuration: configuration)
-        webView.navigationDelegate = self
-        webView.scrollView.decelerationRate = UIScrollViewDecelerationRateNormal
-        return webView
-    }()
-
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-
-        webView.frame = CGRect(origin: .zero, size: self.frame.size)
-        webView.autoresizingMask = [.flexibleWidth, .flexibleHeight]
-        addSubview(webView)
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    func render(html: String, baseURL: URL?) {
-        webView.loadHTMLString(html, baseURL: baseURL)
-    }
-
-    /**
-     Scrolls past `fractionalOffset` of the render view's content size.
-     
-     `scrollToFractionalOffset(_:)` works whenever directly setting the scroll view's `contentOffset` works, but in addition it may work even when the scroll view's `contentSize` is zero.
-     
-     - Seealso: `UIScrollView.fractionalContentOffset`.
-     */
-    func scrollToFractionalOffset(_ fractionalOffset: CGPoint) {
-        let js = "window.scrollTo(document.body.scrollWidth * \(fractionalOffset.x), "
-            + "document.body.scrollHeight * \(fractionalOffset.y))"
-        webView.evaluateJavaScript(js, completionHandler: { result, error in
-            if let error = error {
-                Log.e("error attempting to scroll: \(error)")
-            }
-        })
-    }
-}
-
-extension RenderView: WKNavigationDelegate {
-    func webView(_ webView: WKWebView, decidePolicyFor navigationAction: WKNavigationAction, decisionHandler: @escaping (WKNavigationActionPolicy) -> Void) {
-        guard navigationAction.navigationType == .linkActivated
-            || navigationAction.isAttemptingToHijackWebView
-            || navigationAction.targetFrame == nil else
-        {
-            return decisionHandler(.allow)
-        }
-
-        guard let url = navigationAction.request.url else {
-            return decisionHandler(.allow)
-        }
-
-        decisionHandler(.cancel)
-
-        delegate?.didTapLink(to: url, in: self)
-    }
-}
-
-/**
- A message that can be sent from the render view's JavaScript.
-
- This allows communication from the web view to the native side. For the reverse direction, use methods on `WKWebView`.
- 
- - Seealso: `WKWebView.evaluateJavaScript(_:completionHandler:)`
- */
-protocol RenderViewJavaScriptMessage {
-
-    /// The name of the message. JavaScript can send this message by calling `window.webkit.messageHandlers.messageName.postMessage`, replacing `messageName` with the value returned here.
-    static var messageName: String { get }
-
-    /// - Returns: `nil` if the required message body couldn't be read in `message`.
-    init?(_ message: WKScriptMessage)
-}
-
-extension RenderView: WKScriptMessageHandler {
-    func registerJavaScriptMessage(_ messageType: RenderViewJavaScriptMessage.Type) {
-        registeredJavaScriptMessages.append(messageType)
-
-        webView.configuration.userContentController.add(ScriptMessageHandlerWeakTrampoline(self), name: messageType.messageName)
-    }
-
-    func unregisterJavaScriptMessage(_ messageType: RenderViewJavaScriptMessage.Type) {
-        guard let i = registeredJavaScriptMessages.index(where: { $0 == messageType }) else {
-                return
-        }
-
-        let messageType = registeredJavaScriptMessages.remove(at: i)
-        webView.configuration.userContentController.removeScriptMessageHandler(forName: messageType.messageName)
-    }
-
-    func userContentController(_ userContentController: WKUserContentController, didReceive message: WKScriptMessage) {
-        Log.d("received message from JavaScript: \(message.name)")
-
-        let messageType = registeredJavaScriptMessages.first { $0.messageName == message.name }
-        if let message = messageType?.init(message) {
-            delegate?.didReceive(javaScriptMessage: message, in: self)
-        }
-        else {
-            Log.w("ignoring unexpected message from JavaScript: \(message.name). Did you forget to register a JavaScript message with the RenderView?")
-        }
-    }
-
-    /**
-     Messages that are already present in `RenderView.js` and immediately available to be registered in a `RenderView`. You can add your own messages by conforming to `RenderViewJavaScriptMessage` and registering your message type with `RenderView`.
-     
-     - Seealso: `RenderViewJavaScriptMessage`
-     - Seealso: `RenderView.registerJavaScriptMessage(_:)`
-     */
-    enum JavaScriptMessage {
-
-        /// Sent from the web view once the document has more or less loaded (`DOMContentLoaded`).
-        struct DidRender: RenderViewJavaScriptMessage {
-            static let messageName = "didRender"
-
-            init?(_ message: WKScriptMessage) {
-                assert(message.name == DidRender.messageName)
-            }
-        }
-
-        /// Sent from the web view when the user taps the header in a post.
-        struct DidTapAuthorHeader: RenderViewJavaScriptMessage {
-            static let messageName = "didTapAuthorHeader"
-
-            /// The frame of the tapped header, in the render view's scroll view's coordinate system.
-            let frame: CGRect
-
-            /// The index of the tapped post, where `0` is the first post in the render view.
-            let postIndex: Int
-
-            init?(_ message: WKScriptMessage) {
-                assert(message.name == DidTapAuthorHeader.messageName)
-                
-                guard
-                    let body = message.body as? [String: Any],
-                    let frame = body["frame"] as? [String: Double],
-                    let x = frame["x"],
-                    let y = frame["y"],
-                    let width = frame["width"],
-                    let height = frame["height"],
-                    let postIndex = body["postIndex"] as? Int
-                    else { return nil }
-
-                self.frame = CGRect(x: x, y: y, width: width, height: height)
-                self.postIndex = postIndex
-            }
-        }
-
-        /// Sent from the web view when the user taps the ⋯ button in a post.
-        struct DidTapPostActionButton: RenderViewJavaScriptMessage {
-            static let messageName = "didTapPostActionButton"
-
-            /// The frame of the tapped button, in the render view's scroll view's coordinate system.
-            let frame: CGRect
-
-            /// The index of the tapped post, where `0` is the first post in the render view.
-            let postIndex: Int
-
-            init?(_ message: WKScriptMessage) {
-                assert(message.name == DidTapPostActionButton.messageName)
-
-                guard
-                    let body = message.body as? [String: Any],
-                    let frame = body["frame"] as? [String: Double],
-                    let x = frame["x"],
-                    let y = frame["y"],
-                    let width = frame["width"],
-                    let height = frame["height"],
-                    let postIndex = body["postIndex"] as? Int
-                    else { return nil }
-
-                self.frame = CGRect(x: x, y: y, width: width, height: height)
-                self.postIndex = postIndex
-            }
-        }
-    }
-}
-
-protocol RenderViewDelegate: class {
-    func didReceive(javaScriptMessage: RenderViewJavaScriptMessage, in view: RenderView)
-    func didTapLink(to url: URL, in view: RenderView)
-}
-
-final class Logger {
-    var level: Level
-    private let name: String
-
-    enum Level {
-        case debug, info, warning, error
-
-        fileprivate var abbreviation: String {
-            switch self {
-            case .debug:
-                return "DEBUG"
-            case .info:
-                return "INFO"
-            case .warning:
-                return "WARN"
-            case .error:
-                return "ERROR"
-            }
-        }
-
-        static func >= (lhs: Level, rhs: Level) -> Bool {
-            switch (lhs, rhs) {
-                case (.debug, .debug),
-                     (.info, .debug), (.info, .info),
-                     (.warning, .debug), (.warning, .info), (.warning, .warning),
-                     (.error, _):
-                return true
-
-            case (.debug, _), (.info, _), (.warning, _):
-                return false
-            }
-        }
-    }
-
-    private init(name: String, level: Level = .info) {
-        self.name = name
-        self.level = level
-    }
-
-    private static var loggers: [String: Logger] = [:]
-
-    static func get(_ name: String? = nil, level: Level = .info, file: String = #file) -> Logger {
-        let name = name
-            ?? file.components(separatedBy: "/").last?.components(separatedBy: ".").dropLast().last
-            ?? ""
-
-        if let logger = loggers[name] {
-            return logger
-        }
-
-        let logger = Logger(name: name, level: level)
-        loggers[name] = logger
-
-        return logger
-    }
-
-    private func log(level: Level, message: () -> String, file: String, line: Int) {
-        guard level >= self.level else { return }
-        print("[\(name)] \(file):L\(line):\(level.abbreviation): \(message())")
-    }
-
-    func d(_ message: @autoclosure () -> String, file: String = #file, line: Int = #line) {
-        log(level: .debug, message: message, file: file, line: line)
-    }
-
-    func i(_ message: @autoclosure () -> String, file: String = #file, line: Int = #line) {
-        log(level: .info, message: message, file: file, line: line)
-    }
-
-    func w(_ message: @autoclosure () -> String, file: String = #file, line: Int = #line) {
-        log(level: .warning, message: message, file: file, line: line)
-    }
-
-    func e(_ message: @autoclosure () -> String, file: String = #file, line: Int = #line) {
-        log(level: .error, message: message, file: file, line: line)
-    }
-}
-
-/// WKUserContentController takes a strong reference to its script handlers. That makes a retain cycle when your script handler has a strong reference to the web view. The trampoline breaks the cycle.
-final class ScriptMessageHandlerWeakTrampoline: NSObject, WKScriptMessageHandler {
-    private weak var target: WKScriptMessageHandler?
-
-    init(_ target: WKScriptMessageHandler) {
-        self.target = target
-    }
-
-    func userContentController(_ userContentController: WKUserContentController, didReceive message: WKScriptMessage) {
-        target?.userContentController(userContentController, didReceive: message)
-    }
-}
-
-extension WKNavigationAction {
-    /**
-     Whether the navigation appears to be an embed trying to navigate away.
-     
-     Some embeds (Twitter, YouTube) try to navigate somewhere when the user taps a thing, and for whatever reason it doesn't get counted as a click event.
-     */
-    var isAttemptingToHijackWebView: Bool {
-        guard case .other = navigationType else { return false }
-
-        // TODO: worth considering `targetFrame` and/or `isMainFrame`?
-
-        guard let url = request.url, let host = url.host?.lowercased() else { return false }
-        if host.hasSuffix("www.youtube.com"), url.path.lowercased().hasPrefix("/watch") {
-            return true
-        }
-        else if
-            host.hasSuffix("twitter.com"),
-            let thirdComponent = url.pathComponents.dropFirst(2).first,
-            thirdComponent.lowercased() == "status"
-        {
-            return true
-        }
-        else {
-            return false
-        }
-    }
-}
-
-final class ManagedObjectObserver {
-    private let didChange: (Change) -> Void
-    private var token: NSObjectProtocol?
-
-    enum Change {
-        case delete, update
-    }
-
-    init?(object: NSManagedObject, didChange: @escaping (Change) -> Void) {
-        guard let context = object.managedObjectContext else { return nil }
-
-        self.didChange = didChange
-
-        token = NotificationCenter.default.addObserver(forName: .NSManagedObjectContextObjectsDidChange, object: context, queue: nil) { [unowned self] notification in
-            let notification = ContextObjectsDidChangeNotification(notification)
-            guard let change = Change(object: object, notification: notification) else { return }
-            self.didChange(change)
-        }
-    }
-
-    deinit {
-        if let token = token {
-            NotificationCenter.default.removeObserver(token)
-        }
-    }
-}
-
-struct ContextObjectsDidChangeNotification {
-    private let notification: Notification
-
-    init(_ notification: Notification) {
-        assert(notification.name == .NSManagedObjectContextObjectsDidChange)
-        self.notification = notification
-    }
-
-    private func objects(forKey key: String) -> Set<NSManagedObject> {
-        return notification.userInfo?[key] as? Set<NSManagedObject> ?? []
-    }
-
-    var deletedObjects: Set<NSManagedObject> {
-        return objects(forKey: NSDeletedObjectsKey)
-    }
-
-    var didInvalidateAllObjects: Bool {
-        return notification.userInfo?[NSInvalidatedAllObjectsKey] != nil
-    }
-
-    var invalidatedObjects: Set<NSManagedObject> {
-        return objects(forKey: NSInvalidatedObjectsKey)
-    }
-
-    var refreshedObjects: Set<NSManagedObject> {
-        return objects(forKey: NSRefreshedObjectsKey)
-    }
-
-    var updatedObjects: Set<NSManagedObject> {
-        return objects(forKey: NSUpdatedObjectsKey)
-    }
-}
-
-private extension ManagedObjectObserver.Change {
-    init?(object: NSManagedObject, notification: ContextObjectsDidChangeNotification) {
-        if notification.didInvalidateAllObjects
-            || notification.invalidatedObjects.contains(object)
-            || notification.deletedObjects.contains(object)
-        {
-            self = .delete
-        }
-        else if notification.refreshedObjects.contains(object) || notification.updatedObjects.contains(object) {
-            self = .update
-        }
-        else {
-            return nil
-        }
-    }
-}
-
-infix operator !!: NilCoalescingPrecedence
-
-extension Optional {
-    /// Performs a forced unwrap operation, returning
-    /// the wrapped value of an `Optional` instance
-    /// or performing a `fatalError` with the string
-    /// on the rhs of the operator.
-    ///
-    /// Forced unwrapping unwraps the left-hand side
-    /// if it has a value or errors if it does not.
-    /// The result of a successful operation will
-    /// be the same type as the wrapped value of its
-    /// left-hand side argument.
-    ///
-    /// This operator uses short-circuit evaluation:
-    /// The `optional` lhs is checked first, and the
-    /// `fatalError` is called only if the left hand
-    /// side is nil. For example:
-    ///
-    ///    guard !lastItem.isEmpty else { return }
-    ///    let lastItem = array.last !! "Array guaranteed to be non-empty because..."
-    ///
-    ///    let willFail = [].last !! "Array should have been guaranteed to be non-empty because..."
-    ///
-    ///
-    /// In this example, `lastItem` is assigned the last value
-    /// in `array` because the array is guaranteed to be non-empty.
-    /// `willFail` is never assigned as the last item in an empty array is nil.
-    /// - Parameters:
-    ///   - optional: An optional value.
-    ///   - message: A message to emit via `fatalError` after
-    ///     failing to unwrap the optional.
-    static func !!(optional: Optional, errorMessage: @autoclosure () -> String) -> Wrapped {
-        if let value = optional { return value }
-        fatalError(errorMessage())
     }
 }
